compilation error fixed for lat/lon definition in pdaf1_1. fOr the fix lat/lon changes to real(r8) type.

+++ bldsva/intf_DA/pdaf1_1/model/enkf_clm_mod.F90
@@ -421,8 +421,8 @@ module enkf_clm_mod
     integer :: ni, nj, ii, jj, kk, cid, ier, ncells, nlunits, &
                ncols, npfts, count
     real :: minlon, minlat, maxlon, maxlat
-    real, pointer :: lon(:)
-    real, pointer :: lat(:)
+    real(r8), pointer :: lon(:)
+    real(r8), pointer :: lat(:)
     integer :: begg, endg   ! per-proc gridcell ending gridcell indices
 
     lon   => alatlon%lonc
